refactor options.js and content.js
