+++ lib/content/catPhoto.js
@@ -1,0 +1,3 @@
+export default function catPhoto(n, w, h) {
+  return `https://placekitten.com/${w}/${h}?image=${n}`;
+}
