+++ lib/content/selectors.js
@@ -1,0 +1,37 @@
+export const catWords = [
+  'meooow',
+  'meow',
+  'scratching post',
+  'naww',
+  'purrr',
+  'meeeeow',
+  'kibbles',
+  'cuddle',
+  'snuggle',
+  'purr',
+  'nyaaa',
+  'nyan',
+];
+
+// Selectors.
+export const containerSelector = [
+  '.js-comment-container',
+  '.issue-meta',
+  '.gh-header-meta',
+].join(', ');
+
+export const authorSelector = [
+  '.author',
+  '.author-name > a',
+  '.opened-by > a',
+].join(', ');
+
+export const avatarSelector = [
+  '.avatar',
+  '.timeline-comment-avatar',
+].join(', ');
+
+export const commentSelector = [
+  '.comment-body > p',
+  '.comment-body .email-fragment',
+].join(', ');
