+++ lib/options/elements.js
@@ -1,0 +1,5 @@
+export const themeSelect = document.getElementById('theme');
+export const replaceTextInput = document.getElementById('replace-text');
+export const form = document.getElementById('add-form');
+export const userInput = document.getElementById('user-input');
+export const userList = document.getElementById('user-list');
